always force include rules with grid-area property
if a grid parent using grid-template-areas is critical (f.e. the body), thus appearing in the critical css,
direct children who are supposed to be placed correctly via grid-area (f.e. as a page footer)
can appear incorrectly in the critical viewport if they are not also included.

This change essentially force includes them.
It will make critical css files larger, in vain for most users,
and it takes some time to make this extra check.

The exact extent of the increase in time and size will be researched.